add update method in user manager

+++ src/main/java/com/codecool/controllers/UserManager.java
@@ -1,6 +1,7 @@
 package com.codecool.controllers;
 
 import com.codecool.dao.UserDao;
+import com.codecool.enums.Role;
 import com.codecool.enums.UserInfo;
 import com.codecool.model.Manager;
 import com.codecool.model.User;
@@ -34,7 +35,7 @@ public class UserManager extends Manager {
                     delete();
                     break;
                 case 3:
-                    //do something
+                    update();
                     break;
                 case 4:
                     isRunning = false;
@@ -51,14 +52,17 @@ public class UserManager extends Manager {
     protected void delete() {
         mainView.println("Enter id of user to be removed:");
         User user = userDao.getUserWithUserID(mainView.getIntegerInput());
-        System.out.printf("Current number of records: %d%n", userDao.getNumberOfRecords());
-        System.out.printf("Are you sure you really want to remove user: %s, %s? [Y/N]%n",
-                                                    user.getFirstName(), user.getLastName());
-        if (mainView.getStringInput().equalsIgnoreCase("y")) {
-            userDao.delete(user);
-            mainView.println("\nUser has been removed from database");
+        if (user != null && user.getRoleID() != Role.ADMIN.getRoleID()) {
             System.out.printf("Current number of records: %d%n", userDao.getNumberOfRecords());
-            mainView.pressEnterToContinue("Press enter to go back");
+            System.out.printf("Are you sure you want to remove user: %s, %s? [Y/N]%n", user.getLastName(), user.getFirstName());
+            if (mainView.getStringInput().equalsIgnoreCase("y")) {
+                userDao.delete(user);
+                mainView.println("\nUser has been removed from database");
+                System.out.printf("%nCurrent number of records: %d%n", userDao.getNumberOfRecords());
+                mainView.pressEnterToContinue("\nPress enter to go back");
+            }
+        } else {
+            System.out.println("\nYou cannot delete this user or user doesn't exist\n");
         }
 
     }
@@ -66,12 +70,7 @@ public class UserManager extends Manager {
     @Override
     protected void add() {
         User user = enterUserData();
-        if (user.getRoleID() == 3) {
-            userDao.addCustomerUser(user);
-        } else {
-            userDao.addOtherUser(user);
-        }
-
+        userDao.addUser(user, user.getRoleID());
         user = userDao.validateUser(user);
         System.out.println("\nUser has been created:\n");
         System.out.println(user);
@@ -80,7 +79,27 @@ public class UserManager extends Manager {
 
     @Override
     protected void update() {
-
+        mainView.println("Enter id of user to be updated:");
+        User user = userDao.getUserWithUserID(mainView.getIntegerInput());
+        if (user == null) {
+            System.out.println("\nUser not found in the database\n");
+            return;
+        }
+        String firstName = mainView.readInput("FIRST_NAME", user.getFirstName());
+        String lastName = mainView.readInput("LAST_NAME", user.getLastName());
+        String email = mainView.readInput("EMAIL", user.getEmail());
+        String password = mainView.readInput("PASSWORD", user.getPassword());
+        String address = mainView.readInput("ADDRESS", user.getAddress());
+        String roleID = mainView.readInput("ROLE_ID", String.valueOf(user.getRoleID()));
+        user.setFirstName(firstName);
+        user.setLastName(lastName);
+        user.setEmail(email);
+        user.setPassword(password);
+        user.setAddress(address);
+        user.setRoleID(Integer.parseInt(roleID));
+        userDao.update(user);
+        System.out.println("\nUser has been updated\n");
+        System.out.println(userDao.getUserWithUserID(user.getId()));
     }
 
     public User enterUserData() {
@@ -103,7 +122,7 @@ public class UserManager extends Manager {
         mainView.clearScreen();
         System.out.println("Please enter user's " + field);
         for (int i = 0; i < UserInfo.values().length - 1; i++) {
-            System.out.println(new String[] {"FIRST_NAME", "LAST_NAME", "EMAIL", "ROLEID" } [i] + ": " + answers[i]);
+            System.out.println(new String[] {"FIRST_NAME", "LAST_NAME", "EMAIL", "ROLE_ID" }[i] + ": " + answers[i]);
         }
 
     }
